+++ packages/app/src/app/pages/Dashboard/Components/shared/githubPermissions.tsx
@@ -1,4 +1,10 @@
-import { MessageStripe } from '@codesandbox/components';
+import {
+  Button,
+  Icon,
+  MessageStripe,
+  Stack,
+  Text,
+} from '@codesandbox/components';
 import { useActions, useAppState } from 'app/overmind';
 import React from 'react';
 
@@ -30,5 +36,30 @@ export const RestrictedPublicReposImport: React.FC<{
         Access all repositories
       </MessageStripe.MultiActions>
     </MessageStripe>
+  );
+};
+
+export const RestrictedImportDisclaimer: React.FC = () => {
+  const actions = useActions();
+
+  return (
+    <Stack>
+      <Text>Don&apos;t see all your repositories?</Text>
+      <Button
+        css={{
+          color: '#FFFFFF',
+          cursor: 'pointer',
+          gap: '8px',
+        }}
+        onClick={() =>
+          actions.modalOpened({ modal: 'preferences', itemId: 'integrations' })
+        }
+        variant="link"
+        autoWidth
+      >
+        Review your GitHub permissions
+        <Icon name="external" />
+      </Button>
+    </Stack>
   );
 };
